deployment test 15

+++ api/scripts/application_stop.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
-pkill ts-node &&
-    pkill node
+
+if pgrep ts-node; then pkill ts-node; fi
+
+if pgrep node; then pkill node; fi
+
